Update 10-Nov
Removed "Create Account" page because the option too create an account is in the home page, also there is no place for a "Create Account" button to redirect you to that said page. I've decided to remove it due to that fact.
Other changes:
Created and finished desktop version of following pages.
-Home Page
-Game List Page
-Cart Page
-Checkout Page
-Checkout Success Page
-Added footer with copyright and "The Universe of Games" slogan, styling is done through the styles.css file and not each individual CSS file.