Update 11.Nov
Desktop view of the website is finished, next is too use media query too make the mobile view of the website.

All CSS files have been merged into 1 except for the Media Query which for now will be kept in a separated file.

Corrected previous mistakes such as names, Class instead of ID on a single property (<p>, <h1>, <button> f.eks )
and so on. Another round of "Cleaning" will be done before delivery of the project.